[GlobalOptimization] Move ConvertConv2DToImg2Col pass from Preprocessing
Move the Conv2D to Img2Col transformation pass from Preprocessing to
GlobalOptimization phase. This allows the pass to run after
LinalgQuantizedConvToConvPass, enabling quantized convolutions to
benefit from the img2col transformation.
